Move swapHelpers to the right folder
