+++ src/swap/changehero.ts
@@ -26,7 +26,7 @@ import {
   InvalidCurrencyCodes,
   makeSwapPluginQuote,
   SwapOrder
-} from '../swap-helpers'
+} from '../util/swapHelpers'
 import { convertRequest, getAddress } from '../util/utils'
 import { EdgeSwapRequestPlugin, StringMap } from './types'
