+++ src/swap/defi/thorchain.ts
@@ -29,7 +29,7 @@ import {
   isLikeKind,
   makeSwapPluginQuote,
   SwapOrder
-} from '../../swap-helpers'
+} from '../../util/swapHelpers'
 import {
   convertRequest,
   fetchInfo,
