+++ src/swap/defi/uni-v2-based/plugins/velodrome.ts
@@ -15,7 +15,7 @@ import {
   getMaxSwappable,
   makeSwapPluginQuote,
   SwapOrder
-} from '../../../../swap-helpers'
+} from '../../../../util/swapHelpers'
 import { convertRequest } from '../../../../util/utils'
 import { EdgeSwapRequestPlugin } from '../../../types'
 import VELODROME_V1_ROUTER_ABI from '../../abi/VELODROME_V1_ROUTER_ABI'
