+++ src/swap/godex.ts
@@ -27,7 +27,7 @@ import {
   InvalidCurrencyCodes,
   makeSwapPluginQuote,
   SwapOrder
-} from '../swap-helpers'
+} from '../util/swapHelpers'
 import { convertRequest, getAddress } from '../util/utils'
 import { asNumberString, EdgeSwapRequestPlugin } from './types'
