+++ src/swap/letsexchange.ts
@@ -19,7 +19,7 @@ import {
   InvalidCurrencyCodes,
   makeSwapPluginQuote,
   SwapOrder
-} from '../swap-helpers'
+} from '../util/swapHelpers'
 import { convertRequest, getAddress } from '../util/utils'
 import { asOptionalBlank } from './changenow'
 import { asNumberString, EdgeSwapRequestPlugin } from './types'
