+++ src/swap/transfer.ts
@@ -11,7 +11,7 @@ import {
   getMaxSwappable,
   makeSwapPluginQuote,
   SwapOrder
-} from '../swap-helpers'
+} from '../util/swapHelpers'
 import { convertRequest } from '../util/utils'
 import { EdgeSwapRequestPlugin } from './types'
